pledge-complete/requirements.txt
simple-salesforce
google-cloud-secret-manager